Started parsing latex documents

+++ install.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+# Remove the existing conda environment if it exists
+if conda env list | grep -q "aclready_env"; then
+  conda env remove -n aclready_env
+fi
+
 # Create conda environment if it doesn't exist
 if ! conda env list | grep "aclready_env"; then
   conda env create -f environment.yml
@@ -14,8 +19,12 @@ conda activate aclready_env
 # Install npm dependencies
 npm install --prefix ./aclready
 
-# Create .env file and add .env file to server subfolder
-cat <<EOL > server/.env
+# Check if .env file already exists, if not, create it
+if [ ! -f server/.env ]; then
+  cat <<EOL > server/.env
 TOGETHERAI_API_KEY=your_together_ai_key_here
 OPENAI_API_KEY=your_openai_key_here
 EOL
+else
+  echo ".env file already exists in server subfolder, skipping creation."
+fi
